Crate recursive_array

Macros

recursive_array

a macro for instantiating a recursive array with the given elements.

recursive_array_type_of_size

a macro for getting the type of a generic array with the given item type and size.

Structs

EmptyRecursiveArray

an empty recrusive array.

RecursiveArrayArrayWrapper

a recursive array wrapper which wraps a regular rust array ([T; N]) and allows it to be treated as a recursive array.

RecursiveArrayConcatenation

a recursive array which concatenates 2 recursive arrays.

RecursiveArrayMultiplier

a recursive array which multiplies the given inner recursive array type N times.

RecursiveArraySingleItem

a recursive array with a single item.

Traits

RecursiveArray

a trait which when implemented by some type states that the type’s memory representation can be treated directly as a slice of type T, with a length that is according to the LENGTH constant.