recursive_array

Macro recursive_array_type_of_size

macro_rules! recursive_array_type_of_size {
    ($item_type: ty, $size: expr) => { ... };
}

a macro for getting the type of a generic array with the given item type and size.